requirements.txt
python-dotenv==1.0.0
requests==2.31.0
symbol-sdk-python==3.1.0
websockets==12.0